+++ src/components/StacSource.vue
@@ -1,8 +1,27 @@
 <template>
   <nav class="share">
     <b-button-group>
+      <b-dropdown
+        v-if="canManage" size="sm" variant="outline-primary" tabindex="0"
+      >
+        <template #button-content>
+          <b-icon-gear-fill /><span class="button-label">{{ $t('source.manage') }}</span>
+        </template>
+        <b-dropdown-item v-if="canAddCollections" :to="browserPaths.addCollection">
+          <b-icon-folder-plus /> {{ $t('manage.addCollection') }}
+        </b-dropdown-item>
+        <b-dropdown-item v-if="canAddItems" :to="browserPaths.addItem">
+          <b-icon-file-plus /> {{ $t('manage.addItem') }}
+        </b-dropdown-item>
+        <b-dropdown-item v-if="canEdit" :to="browserPaths.edit">
+          <b-icon-pencil /> {{ $t('manage.edit') }}
+        </b-dropdown-item>
+        <b-dropdown-item v-if="canDelete" @click="confirmDelete = true">
+          <b-icon-trash /> {{ $t('manage.delete') }}
+        </b-dropdown-item>
+      </b-dropdown>
       <b-button
-        v-if="stacUrl" size="sm" variant="outline-primary" id="popover-link-btn"
+        v-if="url" size="sm" variant="outline-primary" id="popover-link-btn"
         :title="$t('source.detailsAboutSource')" tag="a" tabindex="0"
       >
         <b-icon-info-lg /><span class="button-label">{{ $t('source.label') }}</span>
@@ -16,13 +35,13 @@
     </b-button-group>
 
     <b-popover
-      v-if="stacUrl" id="popover-link" class="popover-large" target="popover-link-btn"
+      v-if="url" id="popover-link" class="popover-large" target="popover-link-btn"
       placement="bottom" :title="$t('source.title')" teleport-to="#stac-browser" strategy="fixed"
       click focus :boundary-padding="10"
       v-model="popoverLinkVisible"
     >
       <template #default>
-        <template v-if="stac">
+        <template v-if="data">
           <b-row v-if="stacId" class="stac-id">
             <b-col cols="4">{{ $t('source.id') }}</b-col>
             <b-col>
@@ -37,12 +56,12 @@
           <b-row class="stac-valid">
             <b-col cols="4">{{ $t('source.valid') }}</b-col>
             <b-col>
-              <Validation v-if="popoverLinkVisible !== null" :data="stac" />
+              <Validation v-if="popoverLinkVisible !== null" :data="data" />
             </b-col>
           </b-row>
           <hr>
         </template>
-        <Url id="stacUrl" :url="stacUrl" :label="$t('source.locatedAt')" />
+        <Url id="url" :url="url" :label="$t('source.locatedAt')" />
       </template>
     </b-popover>
     <b-popover
@@ -56,23 +75,41 @@
         <SocialSharing :text="sharingMessage" :title="title" :url="browserUrl()" />
       </template>
     </b-popover>
+    <b-modal :title="$t('manage.confirmDeleteTitle')" v-model="confirmDelete">
+      <p>{{ $t('manage.confirmDeleteMessage') }}</p>
+      <p>{{ $t('manage.noUndo') }}</p>
+      <template #footer="{ close }">
+        <b-button variant="danger" :disabled="deleting" @click="deleteThis">
+          <b-spinner v-if="deleting" small />
+          {{ $t('manage.delete') }}
+        </b-button>
+        <b-button variant="secondary" @click="close()">
+          {{ $t('cancel') }}
+        </b-button>
+      </template>
+    </b-modal>
   </nav>
 </template>
 
 <script>
-import { mapState } from 'vuex';
+import { mapGetters, mapState } from 'vuex';
 import { defineAsyncComponent } from 'vue';
+import { BDropdown, BDropdownItem } from 'bootstrap-vue-next';
 
 import Url from './Url.vue';
 import CopyButton from './CopyButton.vue';
 import SocialSharing from './SocialSharing.vue';
+import { getErrorMessage } from '../store/utils.js';
 
 export default {
   name: "StacSource",
   components: {
+    BDropdown,
+    BDropdownItem,
     Url,
     CopyButton,
     SocialSharing,
+    BModal: defineAsyncComponent(() => import('bootstrap-vue-next').then(m => m.BModal)),
     BPopover: defineAsyncComponent(() => import('bootstrap-vue-next').then(m => m.BPopover)),
     Validation: defineAsyncComponent(() => import('./Validation.vue'))
   },
@@ -80,28 +117,24 @@ export default {
     title: {
       type: String,
       required: true
-    },
-    stacUrl: {
-      type: String,
-      default: null
-    },
-    stac: {
-      type: Object,
-      default: null
     }
   },
   data() {
     return {
-      popoverLinkVisible: null // null = not yet opened, true = open, false = closed
+      popoverLinkVisible: null, // null = not yet opened, true = open, false = closed
+      deleting: false,
+      confirmDelete: false
     };
   },
   computed: {
-    ...mapState(['socialSharing']),
+    ...mapState(['data', 'socialSharing', 'url']),
+    ...mapGetters(['toBrowserPath', 'collectionLink', 'parentLink', 'rootLink']),
+    ...mapGetters('manager', ['browserPaths', 'canEdit', 'canDelete', 'canManage', 'canAddCollections', 'canAddItems']),
     stacVersion() {
-      return this.stac?.stac_version;
+      return this.data?.stac_version;
     },
     stacId() {
-      return this.stac?.id;
+      return this.data?.id;
     },
     enableSocialSharing() {
       return Array.isArray(this.socialSharing) && this.socialSharing.length > 0;
@@ -109,11 +142,39 @@ export default {
     sharingMessage() {
       const url = window.location.toString();
       return this.$t('source.share.message', {title: this.title, url: url});
-    }
+    },
   },
   methods: {
     browserUrl() {
       return window.location.toString();
+    },
+    async deleteThis() {
+      this.deleting = true;
+      const link = {
+        href: this.url,
+        method: 'DELETE'
+      };
+      try {
+        await this.$store.dispatch('request', { link });
+        // Remove the deleted entity from the cache
+        this.$store.commit('clear', this.url);
+        const redirect = this.collectionLink || this.parentLink || this.rootLink;
+        let path = '/';
+        if (redirect) {
+          const redirectUrl = redirect.getAbsoluteUrl();
+          // Remove the parent entity from the cache so that its
+          // children are loaded again without the deleted entity
+          this.$store.commit('clear', redirectUrl);
+          path = this.toBrowserPath(redirectUrl);
+        }
+        this.$router.push(path);
+      } catch (error) {
+        const message = getErrorMessage(error, true);
+        this.$store.commit('showGlobalError', { error, message });
+      } finally {
+        this.deleting = false;
+        this.confirmDelete = false;
+      }
     }
   }
 };