+++ CHANGELOG.md
@@ -28,6 +28,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `relationTypes.config.js` to allow configuring link relation types that
   - are specifically STAC and should be used to navigate to and display in STAC Browser
   - should be hidden
+- Added basic support for the STAC API extensions Transactions (for Items) and Collection Transactions
+  - Adds three new config options: `transactions`, `transactionsRequireLogin` and `transactionsRequirePreflight`
 
 ### Changed